+++ src/pages/card_pages/card.py
@@ -8,33 +8,37 @@ def get_card_columns(account, df, number, detailed=True, make_link=True):
     for index, row in df.iterrows():
         body = [html.H6(str(row.card_name) + '\t\t★ ' + str(row.level),
                         className='card-text',
-                        style={'text-align': 'center', 'margin-bottom': '10px'})]
+                        style={'textAlign': 'center', 'marginBottom': '10px'}
+                        )]
 
         if detailed:
             body.append(html.P('Battles (W-L): ' + str(int(row.win)) + '-' + str(int(row.loss)),
                                className='card-text',
-                               style={'text-align': 'center', 'margin-bottom': '5px'}))
+                               style={'textAlign': 'center', 'marginBottom': '5px'}
+                               ))
 
         body.append(html.P('Battle count: ' + str(int(row.battles)),
                            className='card-text',
-                           style={'text-align': 'center', 'margin-bottom': '5px'}))
+                           style={'textAlign': 'center', 'marginBottom': '5px'}
+                           ))
         if detailed:
             body.append(html.P('Win: ' + str(row.win_percentage) + '%',
                                className='card-text',
-                               style={'text-align': 'center', 'margin-bottom': '5px'}))
+                               style={'textAlign': 'center', 'marginBottom': '5px'}
+                               ))
 
         if make_link:
             cards.append(
                 dbc.CardLink(
                     dbc.Card(children=[
-                            dbc.CardImg(src=row.url, top=True, style={'height': '200px', 'object-fit': 'contain'}),
+                            dbc.CardImg(src=row.url, top=True, style={'height': '200px', 'objectFit': 'contain'}),
                             dbc.CardBody(body),
                         ],
                         className='mb-3 card-hover',
                         style={'padding': '10px'},
                     ),
                     href='card?card_id=' + str(row.card_detail_id) + '#account=' + account,
-                    style={'text-decoration': 'none',
+                    style={'textDecoration': 'none',
                            'color': 'inherit'},
                 ),
             )
@@ -42,7 +46,7 @@ def get_card_columns(account, df, number, detailed=True, make_link=True):
             cards.append(
                 dbc.Card(
                     [
-                        dbc.CardImg(src=row.url, top=True, style={'height': '200px', 'object-fit': 'contain'}),
+                        dbc.CardImg(src=row.url, top=True, style={'height': '200px', 'objectFit': 'contain'}),
                         dbc.CardBody(body),
                     ],
                     className='mb-3',