+++ src/pages/filter_pages/filter_ids.py
@@ -1,0 +1,1 @@
+filter_settings = 'global-filter-settings'