+++ src/pages/navigation_pages/navigation_page.py
@@ -8,16 +8,18 @@ from dash_iconify import DashIconify
 
 from main import app
 from src.configuration import progress, config
-from src.pages import main_page, losing_page
-from src.pages.land_pages import land_page
-from src.pages.rating_pages import rating_page
-from src.pages.season_pages import season_page
+from src.pages import main_page
 from src.pages.card_pages import card_page, card_page_filter
 from src.pages.config_pages import config_page
-from src.pages.shared_modules import styles
+from src.pages.filter_pages import filter_ids
+from src.pages.land_pages import land_page
+from src.pages.losing_pages import losing_page
 from src.pages.navigation_pages import nav_ids
 from src.pages.nemesis_pages import nemesis_page
 from src.pages.portfolio_pages import portfolio_page
+from src.pages.rating_pages import rating_page
+from src.pages.season_pages import season_page
+from src.pages.shared_modules import styles
 from src.utils import store_util
 
 SPL_LOGO = 'https://d36mxiodymuqjm.cloudfront.net/website/icons/img_icon_splinterlands.svg'