+++ src/pages/nemesis_pages/nemesis_page_opponent.py
@@ -23,18 +23,18 @@ def update_opponent(filtered_df):
     if not filtered_df.empty:
         match_type = filtered_df.match_type.value_counts()
         if not match_type.empty:
-            result_layout.append(html.H5("Battles", style={'margin-top': '20px'}))
+            result_layout.append(html.H5("Battles", style={'marginTop': '20px'}))
             for index, value in match_type.items():
                 col_result = []
                 win = filtered_df.loc[(filtered_df.match_type == index) &
                                       (filtered_df.result == 'win')].result.count()
                 loss = filtered_df.loc[(filtered_df.match_type == index) &
                                        (filtered_df.result == 'loss')].result.count()
-                col_result.append(html.H6(str(index) + ':', style={'margin-top': '10px', 'margin-bottom': '0px'}))
-                col_result.append(html.P('Encounters: ' + str(value), style={'margin-bottom': '0px'}))
-                col_result.append(html.P('Win/Loss: ' + str(win) + '-' + str(loss),style={'margin-bottom': '0px'}))
+                col_result.append(html.H6(str(index) + ':', style={'marginTop': '10px', 'marginBottom': '0px'}))
+                col_result.append(html.P('Encounters: ' + str(value), style={'marginBottom': '0px'}))
+                col_result.append(html.P('Win/Loss: ' + str(win) + '-' + str(loss),style={'marginBottom': '0px'}))
                 col_result.append(html.P('Win pct : ' + str(round(win/(win+loss)*100, 2)) + '%',
-                                         style={'margin-bottom': '0px'}))
+                                         style={'marginBottom': '0px'}))
                 result_layout.append(dbc.Col(children=col_result))
 
     return result_layout